scripts/benchmark/requirements.txt
docker==7.1.0
redis==5.0.1
supabase==2.3.0
psutil==5.9.8
rich==13.7.0
click==8.1.7
plotly==5.18.0
jinja2==3.1.2
httpx==0.27.0
python-dotenv==1.0.0
